+++ packages/contentful--create-contentful-app/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.1](https://github.com/contentful/create-contentful-app/compare/@contentful/create-contentful-app@1.9.0...@contentful/create-contentful-app@1.9.1) (2023-08-23)
+
+### Bug Fixes
+
+- **deps:** bump contentful-management from 10.39.2 to 10.40.0 ([#1647](https://github.com/contentful/create-contentful-app/issues/1647)) ([966c358](https://github.com/contentful/create-contentful-app/commit/966c358662ca2c1793ebb2c5eadd28a616c72eb3))
+
 # [1.9.0](https://github.com/contentful/create-contentful-app/compare/@contentful/create-contentful-app@1.8.10...@contentful/create-contentful-app@1.9.0) (2023-08-18)
 
 ### Bug Fixes